+++ src/main/java/com/example/MikoEvents/event/EventRepository.java
@@ -2,6 +2,6 @@ package com.example.MikoEvents.event;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 
 }
