+++ src/main/java/com/example/MikoEvents/organizer/OrganizerRepository.java
@@ -2,6 +2,6 @@ package com.example.MikoEvents.organizer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizerRepository extends JpaRepository<Organizer, Integer> {
+public interface OrganizerRepository extends JpaRepository<Organizer, Long> {
 
 }
