+++ src/main/java/com/example/MikoEvents/registration/RegistrationRepository.java
@@ -2,6 +2,6 @@ package com.example.MikoEvents.registration;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegistrationRepository extends JpaRepository<Registration, Integer> {
+public interface RegistrationRepository extends JpaRepository<Registration, Long> {
 
 }
